Add ListNameAdapter and other

+++ app/src/main/java/by/tigertosh/shoppinglist/base/BaseViewModel.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import by.tigertosh.shoppinglist.database.MainDataBase
 import by.tigertosh.shoppinglist.entities.NoteItem
+import by.tigertosh.shoppinglist.entities.ShoppingListItem
 import by.tigertosh.shoppinglist.entities.ShoppingListName
 import kotlinx.coroutines.launch
 
@@ -15,9 +16,17 @@ class BaseViewModel(database: MainDataBase) : ViewModel() {
     val allNotes: LiveData<List<NoteItem>> = dao.getAllNotes().asLiveData()
     val allShopListNames: LiveData<List<ShoppingListName>> = dao.getAllShopListName().asLiveData()
 
+    fun getAllShopListItem(id: Int): LiveData<List<ShoppingListItem>> =
+        dao.getAllShopListItem(id).asLiveData()
+
     fun insertNote(note: NoteItem) = viewModelScope.launch {
         dao.insertNote(note)
     }
+
+    fun insertShopListItem(listItem: ShoppingListItem) = viewModelScope.launch {
+        dao.insertShopListItem(listItem)
+    }
+
 
     fun insertShopListName(name: ShoppingListName) = viewModelScope.launch {
         dao.insertShopListName(name)
@@ -31,13 +40,18 @@ class BaseViewModel(database: MainDataBase) : ViewModel() {
         dao.updateShoppingListName(listName)
     }
 
+    fun updateListItem(item: ShoppingListItem) = viewModelScope.launch {
+        dao.updateListItem(item)
+    }
+
 
     fun deleteNote(id: Int) = viewModelScope.launch {
         dao.deleteNote(id)
     }
 
-    fun deleteShoppingListName(id: Int) = viewModelScope.launch {
-        dao.deleteShoppingListName(id)
+    fun deleteShoppingList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+        if (deleteList) dao.deleteShoppingListName(id)
+        dao.deleteShopItemByListId(id)
     }
